% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dps}
\alias{dps}
\title{dps}
\format{A data frame with seven variables: \code{X1}, \code{day},
  \code{plate}, \code{sampletype}, \code{sampleID}, \code{gender}, and \code{result}.}
\usage{
dps
}
\description{
example data set simulated according to example 2 in Schaarschmidt et al. (submitted)
}
\details{
120 specimens (una,mbiguously distinguished by the variable sampleID)

technical replications invlove 3 days, 6 plates within each day
the combination of the variables day and plate provides a unique ID variable of the technical 'plate'
variable sampletype distinguishes negative controls ('NC'), unspiked biological specimens ('untreated'), spiked biological specimens ('treated'), and positive controls

normalize: log-transform and substract mean
model structure: days crossed with biological samples (ignoring plate effect within day)
}
\keyword{datasets}

